Split the drawable's transformation logic into its own class for reuse purposes
